Update: auto update

+++ testSignalRInSpring/my-webapp/src/main/java/com/example/restservice/Board/Command.java
@@ -6,13 +6,13 @@ public class Command {
 
     public static class Pair {
 
-        public Pair(double first, double second) {
+        public Pair(int first, int second) {
             this.first = first;
             this.second = second;
         }
 
-        public double first;
-        public double second;
+        public int first;
+        public int second;
     }
 
     public Command(String userId, long timestamp) {
